This is a draft pull request of the point transfer/rasterization framework and spherical, ZB and trilinear kernels. The final commit adds a dummy Houdini SOP which can be used to test the surfacing methods in Houdini, but I anticipate this to be left out of a final PR. Note that this PR also includes the PointStatistics #997 work.

Below are some preliminary timings. More data to be added (timings, images of results, surface comparisons etc).
**Input**: https://artifacts.aswf.io/io/aswf/openvdb/models/waterfall_points.vdb/1.0.0/waterfall_points.vdb-1.0.0.zip

----------------------------------------------------------------------------------------------------------------------------------------------

`tools::rasterizeSpheres()` vs `tools::ParticlesToLevelSet` (From Particle SOP)
No attribute transfers, no pscale on points (reads from parameter)

Target Voxel Size 0.1, Target  Half Band 3 
Input pscale | Old Time (ms) | New Time (ms) | Speedup
-- | -- | -- | -- 
0.1 | 2588.47 | 1764.82 | x1.467 
0.2 | 4435.08 | 2524.45 | x1.757 
0.5 | 17639.2 | 6946.8 | x2.539 
1.0 | 57238.1 | 16061.5 | x3.567

Input  pscale 0.1, Target  Half Band 3 
Target Voxel Size | Old Time (ms) | New Time (ms) | Speedup
-- | -- | -- | -- 
0.01 | 53369.9 | 25959.1 | x2.056 
0.1  | 2588.47 | 1764.82 | x1.467  
1.0  | 4260.55 | 2801.18 | x1.521 
5.0  | 8869.62 | 4194.94 | x2.114

Target Voxel Size 0.1, Input pscale 0.1,
Target Half Band | Old Time (ms) | New Time (ms) | Speedup
-- | -- | -- | -- 
1 | 739.048 | 479.244 | x1.542 
3 | 2588.47 | 1764.82 |  x1.467 
10 | 37725.9 | 19165.4 | x1.968 
15 | 118521 | 51222.2 |  x2.314

----------------------------------------------------------------------------------------------------------------------------------------------

`tools::rasterizeSpheres()` vs `tools::ParticlesToLevelSet` (VDB From Particle SOP)
6 attributes to transfer `v@v` (varying), `v@Cd` (varying), `int@id` (varying), `int64@a` (uniform), `float@b` (uniform). `float@c` (varying)
no pscale on points (reads from parameter)

Target Voxel Size 0.1, Target Half Band 3

Input pscale | Old Time (ms) | New Time (ms) | Speedup
-- | -- | -- | -- 
0.1 | 2815.53 | 1659.43 | x1.697 
0.2 | 4818.51 | 2828.60 | x1.703 
0.5 | 16674.90 | 6948.99 |  x2.4 
1   | 66442.40 | 16395.70 | x4.052 

----------------------------------------------------------------------------------------------------------------------------------------------

`tools::rasterizeSpheres()` vs `tools::ParticlesToLevelSet` (VDB From Particle SOP)
pscale on points (uniform and varying)

// more to document / upload

----------------------------------------------------------------------------------------------------------------------------------------------

`tools::rasterizeSpheres()` vs `tools::rasterizeSmoothSpheres`

This section compares the two new methods against each other and additionally compares `tools::rasterizeSmoothSpheres` against the VDB From Particle Fluid SOP. Note that the latter comparison is not strictly fair; the VDB From Particle Fluid SOP does not perform the exact same functionality as `tools::rasterizeSmoothSpheres`, but in all tests the `Limit Refinement` parameter is on and set to `0`, and the `Resample Input` parameter is `OFF`.

// more to document / upload

Target Voxel Size 0.075, Input Pscale 0.15, Influence Scale (for Smooth Spheres) 3
rasterizeSpheres  |  rasterizeSmoothSpheres
-- | --  
<img src="https://user-images.githubusercontent.com/4256455/111845288-71f5c380-88fc-11eb-82f1-4c57d75bebdc.png" width="500"> | <img src="https://user-images.githubusercontent.com/4256455/111845296-73bf8700-88fc-11eb-950b-02541ee8523c.png" width="500">
2643.64ms | 4361.73ms (x0.6)

Target Voxel Size 0.075, Input Pscale 0.15, Influence Scale (for Smooth Spheres) 3
VDB From Particle Fluid  |  rasterizeSmoothSpheres
-- | --  
<img src="https://user-images.githubusercontent.com/4256455/111845275-702c0000-88fc-11eb-9559-ccb6d51341e5.png" width="500"> | <img src="https://user-images.githubusercontent.com/4256455/111845296-73bf8700-88fc-11eb-950b-02541ee8523c.png" width="500">
5466.03ms | 4361.73ms (x1.253)