@danrbailey 2021-03-17
> @danrbailey, one thing of note might be the commit which modifies some of the point count/mask tools ([e65338d](https://github.com/AcademySoftwareFoundation/openvdb/commit/e65338d0e2d5c996bfa30e76cf42b3bd02ed7514)). It's an isolated/fairly small commit which provides tree only function signatures and faster mask construction. Wanted to provide it in context but happy to split out.

Thanks, yeah I skimmed this bit and it looks good, feel free to split this out and I think I can review this more quickly...