@danrbailey 2021-03-18
Thanks for sharing @Idclip, this is really great to see it all in context. I'm going to have to put aside a fair bit of time to review this. In terms of the core library API at least, one of my main considerations is to try and avoid ending up with _a lot_ of different rasterization methods. It might be useful to me if you can share a more comprehensive list of features and what this rasterization does and does not support. Perhaps based on the all the existing rasterization tools such as VDB Rasterize Points SOP and the Houdini ones such as Volume Rasterize Particles SOP, etc. I'm thinking of features like - varying particle radii, ability to rasterize into frustum grids, velocity blur, arbitrary attribute rasterization support, etc, etc.