@danrbailey 2021-11-23
Seeing as you're working on updating this PR, I wanted to throw out some additional thoughts / considerations on my end. I mentioned some of this in my previous comment so this is not particularly new. 

Rasterization is a massive topic and there are lots of features that we _can_ support. I believe this implementation is not exhaustive in the sense that it does not rasterize to frustum grids, does not support motion blur (see Houdini's Volume Rasterize Attributes SOP), does not support multi-grid rasterization (such as rasterizing thousands of input point grids into a single density grid without having to do them one-by-one), does not support streaming (removes attributes or grids from memory after rasterization completes). To be clear, I'm not suggesting that we should only ever accept an über algorithm that does everything, it seems to me that certain styles of algorithms will be more suitable in some situations than to others. For example, I would think that a scatter-style algorithm would work more efficiently with frustum grids than this type of gather-style algorithm. I just wonder how to organize a world where we might have a bunch of different tools doing some but not all of the features that we associate with rasterization. I'm in one mind to say that all rasterization-related methods should just live in a single header `points/PointRasterize.h` even if there's some overlap and it ups the compile time (but we alleviate that somewhat with explicit instantiation).

One thing that does feel reasonable to me to try and cut down the massive problem space is to refer to points => surface as "surfacing" and points => density/arbitrary grids as "rasterization". So `points/PointRasterizeSDF.h` should be renamed `points/PointSurfacer.h` in my opinion.