@Idclip 2022-01-19
Thanks for your comments @danrbailey - I pretty much agree with what you've said!

> One thing that does feel reasonable to me to try and cut down the massive problem space is to refer to points => surface as "surfacing" and points => density/arbitrary grids as "rasterization". So points/PointRasterizeSDF.h should be renamed points/PointSurfacer.h in my opinion.

The reason I went with `PointRasterizeSDF` was because I also have a `PointRasterizeTrillinear` header ready to PR, so I foresaw a bunch of `PointRasterize<Details>` headers