@danrbailey 2022-02-15
There's a lot to review here, so just for the sake of making some progress with this one, I'm going to add comments as and when I get time and hopefully keep returning to this to get it over the line. 

I'll start with the easy stuff - all the changes to PointMask and PointCount look good, the only thing I would request is that we put a call to the new tree-based methods in the unit tests somewhere just to prevent any subsequent refactoring from unintentionally changing the public API.