@BSalem 2022-03-10
It seems that something was updated which causing rasterizeSpheres() to not be multi-threaded! By applying on the same water_fall.vdb shared here, it's working on a single thread and takes around 17mins! (radius 0.1), I tested on a patched older vdb build (8.0) and it looks like the examples above (it's multi-threaded and taking similar time to process as were shared above).

Any idea about what would be the reason why it's single threaded now?