@Idclip 2022-03-14
> Any idea about what would be the reason why it's single threaded now?

@BSalem thanks for reporting, I remember running into this myself when rebasing these changes. I can't remember exactly what caused it, think it was to do with some threading changes between VDB versions. Currently a bit blocked on proceeding but will investigate this asap.