@Idclip 2022-03-17
@BSalem I beieve I've resolved this - the method now always runs multithreaded and I'll look to expose a control for running it single-threaded in the future. Also started a discussion regarding this issue here: https://lists.aswf.io/g/openvdb-dev/message/1714