@Idclip 2022-03-30
@danrbailey this is good to go. I have more work to do on top of this (such as testing moving the impl out) and we can of course rename and move these files as we progress. Let me know if this looks good to approve.