@RodenLuo 2023-03-02
Hi, I would like to implement a SOP in Houdini for the `rasterizeSmoothSpheres` mentioned in this PR (and adapt it, mainly I would like to have different particle separations for different points when building the SDF). I am freshly new to OpenVDB. I know how to compile an HDK project. But I'm confused about how to compile OpenVDB and use the self-compiled OpenVDB nodes in Houdini.  Could you please kindly guide me or point me to the relevant manual? Thanks!