@Idclip 2023-03-02
Hi @RodenLuo,

The OpenVDB Houdini Plugin can be a bit tricky to compile, but we have some instructions for it here:

https://www.openvdb.org/documentation/doxygen/build.html#buildBuildHou

Also worth mentioning we have the following PR which implements a SOP for this purpose (but is not yet finished, you can use it as a reference if you wish): https://github.com/AcademySoftwareFoundation/openvdb/pull/1390

Nick