@RodenLuo 2023-03-02
Hi Nick, Thanks very much! I was just reading the HDK Doc (and noticed your name in the author filed [here](https://www.sidefx.com/docs/hdk/_point_rasterize_s_d_f_8h.html#a9623e4343ff8ed49dbe3372c9b538517)). It was a headache to wire all the variables correctly. #1390 came right in time! 

Thanks,
Roden